+++ flexmeasures/data/models/planning/storage.py
@@ -38,7 +38,6 @@ from flexmeasures.utils.calculations import (
     integrate_time_series,
 )
 from flexmeasures.utils.time_utils import get_max_planning_horizon
-from flexmeasures.utils.coding_utils import deprecated
 from flexmeasures.utils.time_utils import determine_minimum_resampling_resolution
 from flexmeasures.utils.unit_utils import ur, convert_units
 
@@ -1119,7 +1118,7 @@ class MetaStorageScheduler(Scheduler):
                     site_power_capacity = site_power_capacity.get("sensor", None)
                     if site_power_capacity is None:
                         raise ValueError(
-                            "site-power-capacity attribute is a dict, but has no sensor key."
+                            f"site-power-capacity attribute on asset {asset.id} is a dict, but has no sensor key."
                         )
 
                 power_capacities.append(
@@ -1127,7 +1126,9 @@ class MetaStorageScheduler(Scheduler):
                 )
                 continue
 
-            raise ValueError("Power capacity is not defined in the flex-model.")
+            raise ValueError(
+                f"Power capacity on asset {asset.id} is not defined in the flex-model."
+            )
         return power_capacities
 
     def _ensure_variable_quantity(
@@ -1807,24 +1808,3 @@ def prepend_series(series: pd.Series, value) -> pd.Series:
     # sort index to keep the time ordering
     series = series.sort_index()
     return series.shift(1)
-
-
-#####################
-# TO BE DEPRECATED #
-####################
-@deprecated(build_device_soc_values, "0.14")
-def build_device_soc_targets(
-    targets: list[dict[str, datetime | float]] | pd.Series,
-    soc_at_start: float,
-    start_of_schedule: datetime,
-    end_of_schedule: datetime,
-    resolution: timedelta,
-) -> pd.Series:
-    return build_device_soc_values(
-        targets, soc_at_start, start_of_schedule, end_of_schedule, resolution
-    )
-
-
-StorageScheduler.compute_schedule = deprecated(StorageScheduler.compute, "0.14")(
-    StorageScheduler.compute_schedule
-)